+++ R/Connection.R
@@ -10,6 +10,7 @@ NULL
 #'   Directory authentication token.  Only for use with Microsoft SQL Server and
 #'   with limited support away from the OEM Microsoft driver.
 #' @rdname ConnectionAttributes
+#' @keywords internal
 #' @aliases ConnectionAttributes
 #' @usage NULL
 #' @format NULL
@@ -50,12 +51,21 @@ OdbcConnection <- function(
   attributes = NULL,
   .connection_string = NULL) {
 
-  args <- c(dsn = dsn, driver = driver, server = server, database = database, uid = uid, pwd = pwd, list(...))
-  stopifnot(all(has_names(args)))
+  args <- c(
+    dsn = dsn,
+    driver = driver,
+    server = server,
+    database = database,
+    uid = uid,
+    pwd = pwd,
+    list(...)
+  )
+  check_args(args)
+
   stopifnot(all(has_names(attributes)))
   stopifnot(all(names(attributes) %in% SUPPORTED_CONNECTION_ATTRIBUTES))
 
-  connection_string <- paste0(.connection_string, paste(collapse = ";", sep = "=", names(args), args))
+  connection_string <- paste0(.connection_string, build_connection_string(args))
 
   bigint <- bigint_mappings()[match.arg(bigint, names(bigint_mappings()))]
 
@@ -84,7 +94,38 @@ OdbcConnection <- function(
   res <- new(info$dbms.name, ptr = ptr, quote = quote, info = info, encoding = encoding)
 }
 
+check_args <- function(args) {
+  stopifnot(all(has_names(args)))
+  if (length(args) == 0) {
+    return(args)
+  }
+
+  name_groups <- split(names(args), tolower(names(args)))
+  bad_names <- name_groups[lengths(name_groups) > 1]
+  if (length(bad_names) > 0) {
+    bullets <- vapply(bad_names, paste0, collapse = ", ", FUN.VALUE = character(1))
+
+    abort(
+      c(
+        "After ignoring case, some arguments have the same name:",
+        set_names(bullets, "*")
+      ),
+      call = quote(DBI::dbConnect())
+    )
+  }
+}
+
+build_connection_string <- function(args) {
+  needs_escape <- grepl("[{}(),;?*=!@]", args) &
+    !grepl("^\\{.*\\}$", args) &
+    !vapply(args, inherits, "AsIs", FUN.VALUE = logical(1))
+
+  args[needs_escape] <- paste0("{", args[needs_escape], "}")
+  paste(names(args), args, sep = "=", collapse = ";")
+}
+
 #' @rdname OdbcConnection
+#' @keywords internal
 #' @export
 setClass(
   "OdbcConnection",
@@ -108,9 +149,7 @@ setClass(
 #' to get details on the fields of the table we are writing to.  In particular
 #' the columns `data_type`, `column_size`, and `decimal_digits` are used.  An
 #' implementation is not necessary for [dbWriteTable()] to work.
-#' @param conn OdbcConnection
-#' @param name table we wish to get information on
-#' @param ... additional parameters to methods
+#' @inheritParams odbcConnectionTables
 #'
 #' @seealso The ODBC documentation on [SQLColumns](https://docs.microsoft.com/en-us/sql/odbc/reference/syntax/sqlcolumns-function)
 #' for further details.
@@ -139,7 +178,7 @@ setClass(
 setGeneric(
   "odbcConnectionColumns",
   valueClass = "data.frame",
-  function(conn, name, ...) {
+  function(conn, name, ..., exact = FALSE) {
     standardGeneric("odbcConnectionColumns")
   }
 )
@@ -150,31 +189,41 @@ setGeneric(
 setMethod(
   "odbcConnectionColumns",
   c("OdbcConnection", "Id"),
-  function(conn, name, column_name = NULL) {
-
+  function(conn, name, column_name = NULL, exact = FALSE) {
     odbcConnectionColumns(conn,
       name = id_field(name, "table"),
       catalog_name = id_field(name, "catalog"),
       schema_name = id_field(name, "schema"),
-      column_name = column_name)
+      column_name = column_name,
+      exact = exact
+    )
   }
 )
 
 #' @rdname odbcConnectionColumns
-#' @param catalog_name character catalog where the table is located
-#' @param schema_name character schema where the table is located
 #' @export
 setMethod(
   "odbcConnectionColumns",
   c("OdbcConnection", "character"),
-  function(conn, name, catalog_name = NULL, schema_name = NULL, column_name = NULL) {
+  function(conn,
+           name,
+           catalog_name = NULL,
+           schema_name = NULL,
+           column_name = NULL,
+           exact = FALSE) {
 
-    connection_sql_columns(conn@ptr,
+    if (exact) {
+      schema_name <- escapePattern(schema_name)
+      name <- escapePattern(name)
+      column_name <- escapePattern(column_name)
+    }
+    connection_sql_columns(
+      conn@ptr,
       table_name = name,
       catalog_name = catalog_name,
       schema_name = schema_name,
-      column_name = column_name)
-
+      column_name = column_name
+    )
   }
 )
 
@@ -211,8 +260,14 @@ setMethod(
 #' ( The former also advertises pattern value arguments )
 #'
 #' @param conn OdbcConnection
-#' @param name table we wish to search for
+#' @param name,catalog_name,schema_name Catalog, schema, and table identifiers.
+#'   By default, are interpreted as a ODBC search pattern where `_` and `%` are
+#'   wild cards. Set `exact = TRUE` to match `_` exactly.
 #' @param ... additional parameters to methods
+#' @param exact Set to `TRUE` to escape `_` in identifier names so that it
+#'   matches exactly, rather than matching any single character. `%` always
+#'   matches any number of characters as this is unlikely to appear in a
+#'   table name.
 #'
 #' @seealso The ODBC documentation on [SQLTables](https://docs.microsoft.com/en-us/sql/odbc/reference/syntax/sqlcolumns-function)
 #' for further details.
@@ -222,6 +277,7 @@ setMethod(
 #' - table_schema
 #' - table_name
 #' - table_remarks
+#' @rdname odbcConnectionTables
 setGeneric(
   "odbcConnectionTables",
   valueClass = "data.frame",
@@ -237,32 +293,41 @@ setGeneric(
 setMethod(
   "odbcConnectionTables",
   c("OdbcConnection", "Id"),
-  function(conn, name, table_type = NULL) {
+  function(conn, name, table_type = NULL, exact = FALSE) {
 
-    odbcConnectionTables(conn,
+    odbcConnectionTables(
+      conn,
       name = id_field(name, "table"),
       catalog_name = id_field(name, "catalog"),
       schema_name = id_field(name, "schema"),
-      table_type = table_type)
+      table_type = table_type,
+      exact = exact
+    )
   }
 )
 
 #' @rdname odbcConnectionTables
-#' @param catalog_name character catalog where we wish to query for
-#' available tables
-#' @param schema_name character schema where we wish to query for
-#' available tables.
 setMethod(
   "odbcConnectionTables",
   c("OdbcConnection", "character"),
-  function(conn, name, catalog_name = NULL, schema_name = NULL, table_type = NULL) {
-
-    connection_sql_tables(conn@ptr,
+  function(conn,
+           name,
+           catalog_name = NULL,
+           schema_name = NULL,
+           table_type = NULL,
+           exact = FALSE) {
+    if (exact) {
+      catalog_name <- escapePattern(catalog_name)
+      schema_name <- escapePattern(schema_name)
+      name <- escapePattern(name)
+    }
+    connection_sql_tables(
+      conn@ptr,
       catalog_name = catalog_name,
       schema_name = schema_name,
       table_name = name,
-      table_type = table_type)
-
+      table_type = table_type
+    )
   }
 )
 
@@ -270,22 +335,34 @@ setMethod(
 setMethod(
   "odbcConnectionTables",
   c("OdbcConnection"),
-  function(conn, name = NULL, catalog_name = NULL, schema_name = NULL, table_type = NULL) {
+  function(conn,
+           name = NULL,
+           catalog_name = NULL,
+           schema_name = NULL,
+           table_type = NULL,
+           exact = FALSE) {
 
-    odbcConnectionTables(conn,
+    odbcConnectionTables(
+      conn,
       name = "%",
       catalog_name = catalog_name,
       schema_name = schema_name,
-      table_type = table_type)
-
+      table_type = table_type,
+      exact = exact
+    )
   }
 )
 
 #' @rdname odbcConnectionTables
 setMethod(
   "odbcConnectionTables", c("OdbcConnection", "SQL"),
-  function(conn, name, table_type = NULL) {
-    odbcConnectionTables(conn, dbUnquoteIdentifier(conn, name)[[1]], table_type = table_type)
+  function(conn, name, table_type = NULL, exact = FALSE) {
+    odbcConnectionTables(
+      conn,
+      dbUnquoteIdentifier(conn, name)[[1]],
+      table_type = table_type,
+      exact = exact
+    )
   })
 
 #' odbcConnectionCatalogs
@@ -322,6 +399,7 @@ setMethod(
 #' @param conn OdbcConnection
 #' @param catalog_name Catalog where
 #' we are looking to list schemas.
+#' @keywords internal
 #' @rdname odbcConnectionSchemas
 setGeneric(
   "odbcConnectionSchemas",
@@ -494,45 +572,64 @@ setMethod(
   "dbQuoteIdentifier", c("OdbcConnection", "SQL"),
   getMethod("dbQuoteIdentifier", c("DBIConnection", "SQL"), asNamespace("DBI")))
 
-#' @inherit DBI::dbListTables
-#' @param catalog_name The name of the catalog to return, the default returns all catalogs.
-#' @param schema_name The name of the schema to return, the default returns all schemas.
-#' @param table_name The name of the table to return, the default returns all tables.
+#' List remote tables and fields for an ODBC connection
+#'
+#' `dbListTables()` provides names of remote tables accessible through this
+#' connection; `dbListFields()` provides names of columns within a table.
+#'
+#' @inheritParams DBI::dbListTables
+#' @param catalog_name,schema_name,table_name Catalog, schema, and table names.
+#'
+#'   By default, `catalog_name`, `schema_name` and `table_name` will
+#'   automatically escape underscores to ensure that you match exactly one
+#'   table. If you want to search for multiple tables using wild cards, you
+#'   will need to use `odbcConnectionTables()` directly instead.
+#'
 #' @param table_type The type of the table to return, the default returns all table types.
-#' @aliases dbListTables
-#' @details
-#' \code{\%} can be used as a wildcard in any of the search parameters to
-#'   match 0 or more characters. `_` can be used to match any single character.
-#' @seealso The ODBC documentation on [Pattern Value Arguments](https://docs.microsoft.com/en-us/sql/odbc/reference/develop-app/pattern-value-arguments)
-#'   for further details on the supported syntax.
+#' @returns A character vector of table or field names respectively.
 #' @export
 setMethod(
   "dbListTables", "OdbcConnection",
-  function(conn, catalog_name = NULL, schema_name = NULL, table_name = NULL,
-    table_type = NULL, ...) {
+  function(conn,
+           catalog_name = NULL,
+           schema_name = NULL,
+           table_name = NULL,
+           table_type = NULL,
+           ...) {
 
-    odbcConnectionTables(conn,
+    tables <- odbcConnectionTables(
+      conn,
       name = table_name,
       catalog_name = catalog_name,
       schema_name = schema_name,
-      table_type = table_type)$table_name
+      table_type = table_type,
+      exact = TRUE
+    )
+    tables[["table_name"]]
   })
 
-#' @inherit DBI::dbListFields
+#' @rdname dbListTables-OdbcConnection-method
 #' @inheritParams DBI::dbListFields
-#' @aliases dbListFields
-#' @inheritParams dbListTables,OdbcConnection-method
 #' @param column_name The name of the column to return, the default returns all columns.
-#' @inherit dbListTables,OdbcConnection-method details
 #' @export
 setMethod(
   "dbListFields", c("OdbcConnection", "character"),
-  function(conn, name, catalog_name = NULL, schema_name = NULL, column_name = NULL, ...) {
-    odbcConnectionColumns(conn,
+  function(
+      conn,
+      name,
+      catalog_name = NULL,
+      schema_name = NULL,
+      column_name = NULL,
+      ...
+  ) {
+    cols <- odbcConnectionColumns(
+      conn,
       name = name,
       catalog_name = catalog_name,
       schema_name = schema_name,
-      column_name = column_name)[["name"]]
+      column_name = column_name,
+      exact = TRUE)
+    cols[["name"]]
   })
 
 #' @rdname OdbcConnection
@@ -606,26 +703,83 @@ setMethod(
     invisible(TRUE)
   })
 
-#' List Available ODBC Drivers
+#' List Configured ODBC Drivers
 #'
-#' List the available drivers on your system. See the
-#' [Installation](https://github.com/r-dbi/odbc#installation) section of the
-#' package README for details on how to install drivers for the most common
-#' databases.
+#' @description
 #'
-#' @param keep A character vector of driver names to keep in the results, if
-#'   `NULL` (the default) will keep all drivers.
-#' @param filter A character vector of driver names to filter from the results, if
-#'   `NULL` (the default) will not filter any drivers.
+#' Collect information about the configured driver names. A driver must be both
+#' installed and configured with the driver manager to be included in this list.
+#' Configuring a driver name just sets up a lookup table (e.g. in
+#' `odbcinst.ini`) to allow users to pass only the driver name to [dbConnect()].
+#'
+#' Driver names that are not configured with the driver manager (and thus
+#' do not appear in this function's output) can still be
+#' used in [dbConnect()] by providing a path to a driver directly.
+#'
+#' @param keep,filter A character vector of driver names to keep in or remove
+#'   from the results, respectively. If `NULL`, all driver names will be kept,
+#'   or none will be removed, respectively. The `odbc.drivers_keep` and
+#'   `odbc.drivers_filter` options control the argument defaults.
+#'
+#'   Driver names are first processed with `keep`, then `filter`. Thus, if a
+#'   driver name is in both `keep` and `filter`, it won't appear in output.
+#'
 #' @return A data frame with three columns.
-#' If a given driver does not have any attributes the last two columns will be
-#' `NA`. Drivers can be excluded from being returned by setting the
-#' \code{odbc.drivers.filter} option.
+#'
 #' \describe{
-#'   \item{name}{Name of the driver}
-#'   \item{attribute}{Driver attribute name}
-#'   \item{value}{Driver attribute value}
+#'   \item{name}{Name of the driver. The entries in this column can be
+#'     passed to the `driver` argument of [dbConnect()] (as long as the
+#'     driver accepts the argument).}
+#'   \item{attribute}{Driver attribute name.}
+#'   \item{value}{Driver attribute value.}
 #' }
+#'
+#' If a driver has multiple attributes, there will be one row per attribute,
+#' each with the same driver `name`. If a given driver name does not have any
+#' attributes, the function will return one row with the driver `name`, but
+#' the last two columns will be `NA`.
+#'
+#' @section Configuration:
+#'
+#' This function interfaces with the driver manager to collect information
+#' about the available driver names.
+#'
+#' For **MacOS and Linux**, the odbc package supports the unixODBC driver
+#' manager. unixODBC looks to the `odbcinst.ini` _configuration file_ for
+#' information on driver names. Find the location(s) of your `odbcinst.ini`
+#' file(s) with `odbcinst -j`.
+#'
+#' In this example `odbcinst.ini` file:
+#'
+#' ```
+#' [MySQL Driver]
+#' Driver=/opt/homebrew/Cellar/mysql/8.2.0_1/lib/libmysqlclient.dylib
+#' ```
+#'
+#' Then the driver name is `MySQL Driver`, which will appear in the `name`
+#' column of this function's output. To pass the driver name as the `driver`
+#' argument to [dbConnect()], pass it as a string, like `"MySQL Driver"`.
+#'
+#' **Windows** is [bundled](https://learn.microsoft.com/en-us/sql/odbc/admin/odbc-data-source-administrator)
+#' with an ODBC driver manager.
+#'
+#' In this example, function output would include 1 row: the `name` column
+#' would read `"MySQL Driver"`, `attribute` would be `"Driver"`, and `value`
+#' would give the file path to the driver. Additional key-value pairs
+#' under the driver name would add additional rows with the same `name` entry.
+#'
+#' When a driver is configured with a driver manager, information on the driver
+#' will be automatically passed on to [dbConnect()] when its `driver` argument
+#' is set. For an example, see the same section in the [odbcListDataSources()]
+#' help-file. Instead of configuring driver information with a driver manager,
+#' it is also possible to provide a path to a driver directly to [dbConnect()].
+#'
+#' @seealso
+#' [odbcListDataSources()]
+#'
+#' @examplesIf FALSE
+#' odbcListDrivers()
+#'
 #' @export
 odbcListDrivers <- function(keep = getOption("odbc.drivers_keep"), filter = getOption("odbc.drivers_filter")) {
   res <- list_drivers_()
@@ -645,18 +799,88 @@ odbcListDrivers <- function(keep = getOption("odbc.drivers_keep"), filter = getO
   res
 }
 
-#' List Available Data Source Names
+#' List Configured Data Source Names
 #'
-#' List the available data sources on your system. See the [DSN Configuration
-#' files](https://github.com/r-dbi/odbc#dsn-configuration-files) section of the
-#' package README for details on how to install data sources for the most
-#' common databases.
+#' @description
 #'
-#' @return A data frame with two columns.
+#' Collect information about the available data source names (DSNs). A DSN must
+#' be both installed and configured with the driver manager to be included in
+#' this list. Configuring a DSN just sets up a lookup table (e.g. in
+#' `odbc.ini`) to allow users to pass only the DSN to [dbConnect()].
+#'
+#' DSNs that are not configured with the driver manager can still be
+#' connected to with [dbConnect()] by providing DSN metadata directly.
+#'
+#' @return A data frame with two columns:
 #' \describe{
-#'   \item{name}{Name of the data source}
-#'   \item{description}{Data Source description}
+#'   \item{name}{Name of the data source. The entries in this column can be
+#'     passed to the `dsn` argument of [dbConnect()].}
+#'   \item{description}{Data source description.}
 #' }
+#'
+#' @section Configuration:
+#'
+#' This function interfaces with the driver manager to collect information
+#' about the available data source names.
+#'
+#' For **MacOS and Linux**, the odbc package supports the unixODBC driver
+#' manager. unixODBC looks to the `odbc.ini` _configuration file_ for information
+#' on DSNs. Find the location(s) of your `odbc.ini` file(s) with `odbcinst -j`.
+#'
+#' In this example `odbc.ini` file:
+#'
+#' ```
+#' [MySQL]
+#' Driver = MySQL Driver
+#' Database = test
+#' Server = 127.0.0.1
+#' User = root
+#' password = root
+#' Port = 3306
+#' ```
+#'
+#' ...the data source name is `MySQL`, which will appear in the `name`
+#' column of this function's output. To pass the DSN as the `dsn` argument to
+#' [dbConnect()], pass it as a string, like `"MySQL"`.
+#' `Driver = MySQL Driver` references the driver `name` in [odbcListDrivers()]
+#' output.
+#'
+#' **Windows** is [bundled](https://learn.microsoft.com/en-us/sql/odbc/admin/odbc-data-source-administrator)
+#' with an ODBC driver manager.
+#'
+#' When a DSN is configured with a driver manager, information on the DSN will
+#' be automatically passed on to [dbConnect()] when its `dsn` argument is set.
+#'
+#' For example, with the `MySQL` data source name configured, and the driver
+#' name `MySQL Driver` appearing in [odbcListDrivers()] output, the code:
+#'
+#' ```
+#' con <-
+#'   dbConnect(
+#'     odbc::odbc(),
+#'     Driver = "MySQL Driver",
+#'     Database = "test",
+#'     Server = "127.0.0.1",
+#'     User = "root",
+#'     password = "root",
+#'     Port = 3306
+#'   )
+#' ```
+#'
+#' ...can be written:
+#'
+#' ```
+#' con <- dbConnect(odbc::odbc(), dsn = "MySQL")
+#' ```
+#'
+#' In this case, `dbConnect()` will look up the information defined for `MySQL`
+#' in the driver manager (in our example, `odbc.ini`) and automatically
+#' pass the needed arguments.
+#'
+#' @seealso
+#'
+#' [odbcListDrivers()]
+#'
 #' @export
 odbcListDataSources <- function() {
   list_data_sources_()