+++ R/Table.R
@@ -24,7 +24,7 @@
 #'
 #' dbDisconnect(con)
 #' }
-#' @name odbc-tables
+#' @name DBI-tables
 NULL
 
 odbc_write_table <-
@@ -68,7 +68,7 @@ odbc_write_table <-
     }
 
     fieldDetails <- tryCatch({
-      details <- odbcConnectionColumns(conn, name)
+      details <- odbcConnectionColumns(conn, name, exact = TRUE)
       details$param_index <- match(details$name, names(values))
       details[!is.na(details$param_index) & !is.na(details$data_type), ]
     }, error = function(e) {
@@ -101,7 +101,7 @@ odbc_write_table <-
     invisible(TRUE)
   }
 
-#' @rdname odbc-tables
+#' @rdname DBI-tables
 #' @inheritParams DBI::dbWriteTable
 #' @param overwrite Allow overwriting the destination table. Cannot be
 #'   `TRUE` if `append` is also `TRUE`.
@@ -115,19 +115,19 @@ setMethod(
   "dbWriteTable", c("OdbcConnection", "character", "data.frame"),
   odbc_write_table)
 
-#' @rdname odbc-tables
+#' @rdname DBI-tables
 #' @export
 setMethod(
   "dbWriteTable", c("OdbcConnection", "Id", "data.frame"),
   odbc_write_table)
 
-#' @rdname odbc-tables
+#' @rdname DBI-tables
 #' @export
 setMethod(
   "dbWriteTable", c("OdbcConnection", "SQL", "data.frame"),
   odbc_write_table)
 
-#' @rdname odbc-tables
+#' @rdname DBI-tables
 #' @inheritParams DBI::dbAppendTable
 #' @export
 setMethod("dbAppendTable", "OdbcConnection", function(conn, name, value, ..., row.names = NULL) {
@@ -137,7 +137,7 @@ setMethod("dbAppendTable", "OdbcConnection", function(conn, name, value, ..., ro
   invisible(NA_real_)
 })
 
-#' @rdname odbc-tables
+#' @rdname DBI-methods
 #' @inheritParams DBI::dbReadTable
 #' @export
 setMethod("sqlData", "OdbcConnection", function(con, value, row.names = NA, ...) {
@@ -163,7 +163,7 @@ setMethod("sqlData", "OdbcConnection", function(con, value, row.names = NA, ...)
   value
 })
 
-#' @rdname odbc-tables
+#' @rdname DBI-tables
 #' @inheritParams DBI::sqlCreateTable
 #' @param field.types Additional field types used to override derived types.
 #' @export
@@ -232,6 +232,6 @@ setMethod(
   "dbExistsTable", c("OdbcConnection", "character"),
   function(conn, name, ...) {
     stopifnot(length(name) == 1)
-    df <- odbcConnectionTables(conn, name = name, ...)
+    df <- odbcConnectionTables(conn, name = name, ..., exact = TRUE)
     NROW(df) > 0
   })