+++ R/db.R
@@ -45,6 +45,7 @@ setMethod(
 # Simple class prototype to avoid messages about unknown classes from setMethod
 setClass("Oracle", where = class_cache)
 
+#' @rdname DBI-methods
 setMethod("sqlCreateTable", "Oracle",
   function(con, table, fields, row.names = NA, temporary = FALSE, ..., field.types = NULL) {
     table <- dbQuoteIdentifier(con, table)
@@ -65,9 +66,9 @@ setMethod("sqlCreateTable", "Oracle",
 setMethod(
   "odbcConnectionTables",
   c("Oracle", "character"),
-  function(conn, name, catalog_name = NULL, schema_name = NULL, table_type = NULL) {
+  function(conn, name, catalog_name = NULL, schema_name = NULL, table_type = NULL, exact = FALSE) {
 
-    qTable <- getSelector("object_name", name)
+    qTable <- getSelector("object_name", name, exact)
     if (is.null(schema_name)) {
       query <- paste0(
         " SELECT null AS \"table_catalog\", '", conn@info$username ,"' AS \"table_schema\", object_name AS \"table_name\", object_type AS \"table_type\", null AS \"table_remarks\"",
@@ -75,7 +76,7 @@ setMethod(
         " WHERE 1 = 1 ", qTable,
         " AND ( object_type = 'TABLE' OR object_type = 'VIEW' ) ")
     } else {
-      qSchema <- getSelector("owner", schema_name)
+      qSchema <- getSelector("owner", schema_name, exact)
       query <- paste0(
         " SELECT null AS \"table_catalog\", owner AS \"table_schema\", object_name AS \"table_name\", object_type AS \"table_type\", null AS \"table_remarks\"",
         " FROM all_objects ",
@@ -92,7 +93,7 @@ setMethod(
 setMethod(
   "odbcConnectionColumns",
   c("Oracle", "character"),
-  function(conn, name, catalog_name = NULL, schema_name = NULL, column_name = NULL) {
+  function(conn, name, catalog_name = NULL, schema_name = NULL, column_name = NULL, exact = FALSE) {
 
     query <- ""
     baseSelect <- paste0(
@@ -114,7 +115,7 @@ setMethod(
       decode(data_type,'CHAR',data_length,'VARCHAR2',data_length,'NVARCHAR2',data_length,'NCHAR',data_length, 0) AS \"char_octet_length\",
       column_id AS \"ordinal_position\",
       decode(nullable, 'Y', 1, 'N', 0) AS \"nullable\"")
-    qTable <- getSelector("table_name", name)
+    qTable <- getSelector("table_name", name, exact)
     if (is.null(schema_name)) {
       baseSelect <- gsub("owner AS \"schema_name\"", paste0("'", conn@info$username, "' AS \"schema_name\""), baseSelect);
       query <- paste0(
@@ -122,7 +123,7 @@ setMethod(
          " FROM user_tab_columns ",
          " WHERE 1 = 1 ", qTable );
     } else {
-      qSchema <- getSelector("owner", schema_name)
+      qSchema <- getSelector("owner", schema_name, exact)
       query <- paste0(
         baseSelect,
         " FROM all_tab_columns ",
@@ -136,6 +137,7 @@ setMethod(
 
 setClass("Teradata", where = class_cache)
 
+#' @rdname DBI-methods
 setMethod("sqlCreateTable", "Teradata",
   function(con, table, fields, row.names = NA, temporary = FALSE, ..., field.types = NULL) {
     table <- dbQuoteIdentifier(con, table)
@@ -191,6 +193,7 @@ setMethod(
 
 setClass("HDB", where = class_cache)
 
+#' @rdname DBI-methods
 setMethod("sqlCreateTable", "HDB",
   function(con, table, fields, row.names = NA, temporary = FALSE, ..., field.types = NULL) {
     table <- dbQuoteIdentifier(con, table)
@@ -209,6 +212,7 @@ setMethod("sqlCreateTable", "HDB",
 
 setClass("Hive", where = class_cache)
 
+#' @rdname DBI-methods
 setMethod(
   # only need to override dbQuteString when x is character.
   # DBI:::quote_string just returns x when it is of class SQL, so no need to override that.
@@ -254,6 +258,11 @@ setMethod(
 
 setClass("DB2/AIX64", where = class_cache)
 
+#' @rdname DBI-methods
+# Don't generate a usage because there's some buglet in R CMD check
+# (probably because of the `/` in the class name) which flags a usage
+# without corresponding alias
+#' @usage NULL
 setMethod("sqlCreateTable", "DB2/AIX64",
   function(con, table, fields, row.names = NA, temporary = FALSE, ..., field.types = NULL) {
     table <- dbQuoteIdentifier(con, table)
@@ -270,7 +279,6 @@ setMethod("sqlCreateTable", "DB2/AIX64",
 
 # Microsoft SQL Server ---------------------------------------------------------
 
-#' Simple class prototype to avoid messages about unknown classes from setMethod
 #' @rdname SQLServer
 #' @usage NULL
 setClass("Microsoft SQL Server", where = class_cache)
@@ -290,6 +298,7 @@ setClass("Microsoft SQL Server", where = class_cache)
 #' @docType methods
 #' @inheritParams DBI::dbUnquoteIdentifier
 #' @usage NULL
+#' @keywords internal
 setMethod("dbUnquoteIdentifier", c("Microsoft SQL Server", "SQL"),
   function(conn, x, ...) {
     x <- gsub("(\\[)([^\\.]+?)(\\])", "\\2", x)
@@ -339,7 +348,6 @@ setMethod("isTempTable", c("Microsoft SQL Server", "character"),
 #' Therefore, in all cases query for \code{name___%}.
 #' @rdname SQLServer
 #' @docType methods
-#' @aliases dbExistsTable
 #' @inherit DBI::dbExistsTable
 #' @usage NULL
 setMethod(