+++ tests/testthat/test-SQLite.R
@@ -1,8 +1,10 @@
 test_that("SQLite", {
-  skip_unless_has_test_db({
-    DBItest::make_context(odbc(), list(.connection_string = Sys.getenv("ODBC_CS")),
-      tweaks = DBItest::tweaks(placeholder_pattern = "?", strict_identifier = TRUE), name = "SQLite")
-  })
+  DBItest::make_context(
+    odbc(),
+    test_connection_string("SQLITE"),
+    tweaks = DBItest::tweaks(placeholder_pattern = "?", strict_identifier = TRUE),
+    name = "SQLite"
+  )
 
   DBItest::test_getting_started(c(
       "package_name", # Not an error